refactor: replace heboImage import with ecs in API, Auth, Gateway, and MCP services to standardize initialization process handling
